refactor(main_logic): core.py 拆分为 core/ 包（阶段一：纯搬模块级代码）
上帝文件拆分路线图第一批第 2 刀。main_logic/core.py（10,420 行）→
main_logic/core/ 包，import 路径 main_logic.core 不变（复用 #2148
system_router 门面模式）：

- core/_shared.py：模块级常量、包 logger、语音回声检测、locale 文案、
  _proactive_expected_sid ContextVar、_purge_closed_tool_calls 等纯助手
- core/callback_render.py：agent 回调/语音热切换注入的纯函数渲染族
  （apply_role_placeholders、_build_callback_instruction 等）
- core/notices.py：重要通知缓冲池（队列三全局唯一属主）
- core/__init__.py：LLMSessionManager 类本体原样 + 全量 re-export 门面

LLMSessionManager 刻意留在 __init__.py：类方法的 __globals__ 即门面模块
dict，25+ 测试文件对 main_logic.core.<attr> 的 monkeypatch（
CROSS_MODE_RESTART_WAIT_SECONDS、is_livestream_active、
aload_global_conversation_settings 等）继续原样命中类内消费点。

仅有的两处非纯搬编辑（保持行为不变所需）：
- logger 显式绑定 "main_logic.core"（保住测试断言的
  N.E.K.O.Main.main_logic.core logger 名）
- _STATIC_LOCALES_DIR parents[1]→parents[2]（文件深了一层）

_prominent_notice_seq 故意不 re-export：属主每次 enqueue 重绑该 int，
门面快照必然过期；已核实无外部读者。
